Fix error on first-time apply
This commit applies the Hashicorp-recommended
[workaround](https://github.com/hashicorp/terraform-provider-aws/issues/32025#issuecomment-1599344398)
for the bug where Terraform tries to use ACM certificates before they
are available. We are already using a resource,
`aws_acm_certificate_validation`, that depends on the certificate and
is smart enough to wait for the certificate to be available. That
resource also exposes the certificate ARN, so we can is re-order the
dependency graph to use rely on it instead.

Previous:
Cert -> Verification (works correctly)
Cert -> Domain Name Config (broken)

New:
Cert -> Verification -> Domain Name Config